+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 ## Unreleased
 ### Added
 - Get cluster labels method to InstanceClient class.
+### Fixed
+- Possible password leak in LocalExecutor.
 
 ## [0.4.0][changes-0.4.0] - 2025-12-10
 ### Added